# backend/docs/swagger.json
{
    "components": {
        "schemas": {
...
            "models.CreateTransactionRequest": {
                "properties": {
                    "amount": {
                        "type": "number"
                    },
                    "category": {
                        "type": "string"
                    },
                    "date": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "purpose": {
                        "type": "string"
                    },
                    "source": {
                        "type": "string"
                    },
                    "type": {
                        "$ref": "#/components/schemas/models.TransactionType"
                    },
                    "wishlist_id": {
                        "type": "string"
                    }
                },
                "required": [
                    "amount",
                    "category",
                    "date",
                    "type"
                ],
                "type": "object"
            },
...
            "models.CreateWishlistRequest": {
                "properties": {
                    "category": {
                        "type": "string"
                    },
                    "estimated_cost": {
                        "type": "number"
                    },
                    "frequency": {
                        "$ref": "#/components/schemas/models.Frequency"
                    },
                    "name": {
                        "type": "string"
                    },
                    "notes": {
                        "type": "string"
                    },
                    "priority": {
                        "type": "integer"
                    },
                    "purchased": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "category",
                    "estimated_cost",
                    "name"
                ],
                "type": "object"
            },
...
            "models.TransactionType": {
                "enum": [
                    "income",
                    "expense",
                    "initial_balance"
                ],
                "type": "string",
                "x-enum-varnames": [
                    "Income",
                    "Expense",
                    "InitialBalance"
                ]
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "description": "Заголовок вида `Bearer \u003cjwt\u003e` (получить через `/auth/login`).",
                "in": "header",
                "name": "Authorization",
                "type": "apiKey"
            }
        }
    },
    "info": {
        "contact": {
            "name": "msdnna",
            "url": "https://github.com/msdnna/budget"
        },
        "description": "REST API семейного бюджет-приложения (Go + MongoDB). JWT (Bearer) — защищённые ручки требуют заголовок `Authorization: Bearer \u003ctoken\u003e`, полученный через `POST /api/auth/login`.",
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0"
        },
        "title": "msdnna Budget API",
        "version": "1.0"
    },
...
    "paths": {
...
        "/transactions": {
            "get": {
                "description": "Фильтры комбинируются по И. `categories` — список через запятую (приоритетнее `category`). `include_detailed=true` показывает родителей закрытых detail-requests.",
                "parameters": [
                    {
                        "description": "income|expense|initial_balance",
                        "in": "query",
                        "name": "type",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Имя категории",
                        "in": "query",
                        "name": "category",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "CSV категорий",
                        "in": "query",
                        "name": "categories",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "YYYY-MM-DD",
                        "in": "query",
                        "name": "from",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "YYYY-MM-DD",
                        "in": "query",
                        "name": "to",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Страница (1-based)",
                        "in": "query",
                        "name": "page",
                        "schema": {
                            "default": 1,
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Размер страницы (1-100)",
                        "in": "query",
                        "name": "limit",
                        "schema": {
                            "default": 20,
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Показывать родителей закрытых detail-requests",
                        "in": "query",
                        "name": "include_detailed",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {},
                                    "type": "object"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Список транзакций с пагинацией",
                "tags": [
                    "transactions"
                ]
            },
            "post": {
                "description": "Создаёт доход / расход / начальный баланс. `date` принимается в формате YYYY-MM-DD.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/models.CreateTransactionRequest",
                                        "summary": "body",
                                        "description": "Тело"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Тело",
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.Transaction"
                                }
                            }
                        },
                        "description": "Created"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Создать транзакцию",
                "tags": [
                    "transactions"
                ]
            }
        },
...
        "/wishlist": {
            "get": {
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/models.WishlistItem"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Все wishlist-итемы (без soft-deleted)",
                "tags": [
                    "wishlist"
                ]
            },
            "post": {
                "description": "`frequency=once` — обычная покупка из списка желаний. `monthly|quarterly|yearly` — регулярный платёж (коммуналка, связь и т.п.).",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/models.CreateWishlistRequest",
                                        "summary": "body",
                                        "description": "Тело"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Тело",
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.WishlistItem"
                                }
                            }
                        },
                        "description": "Created"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Создать wishlist-итем (одноразовый или регулярный)",
                "tags": [
                    "wishlist"
                ]
            }
        },
        "/wishlist/{id}": {
            "delete": {
                "parameters": [
                    {
                        "description": "Wishlist ID",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Удалить wishlist-итем (soft-delete)",
                "tags": [
                    "wishlist"
                ]
            },
            "put": {
                "description": "Patch-семантика. `purchased=true` помечает «куплено» (используется UI-кнопкой); транзакция списания создаётся отдельным `POST /transactions`.",
                "parameters": [
                    {
                        "description": "Wishlist ID",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/models.UpdateWishlistRequest",
                                        "summary": "body",
                                        "description": "Поля"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Поля",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.WishlistItem"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Обновить wishlist-итем",
                "tags": [
                    "wishlist"
                ]
            }
        },
...
        "/wishlist/{id}/unlink-period": {
            "post": {
                "description": "Очищает `wishlist_id` у транзакций в текущем месяце/квартале/году (по `frequency`). Для `once` — отвязывает единственную привязанную транзакцию без date-filter. Бэкенд для кнопок «Не куплено» и «Отменить».",
                "parameters": [
                    {
                        "description": "Wishlist ID",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "integer"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Отвязать транзакции wishlist-итема за текущий период",
                "tags": [
                    "wishlist"
                ]
            }
        }
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "url": "/api"
        }
    ]
}